Extension replaced with subclass in order to have stored properties w/o using the Obj-C runtime
