+++ GiniVision/Classes/Networking/Pinning/GiniVision+Pinning.swift
@@ -34,7 +34,7 @@ extension GiniVision {
                                            documentMetadata: GINIDocumentMetadata? = nil,
                                            api: GINIAPIType = .default) -> UIViewController {
         GiniVision.setConfiguration(configuration)
-        let screenCoordinator = GiniScreenAPICoordinator(client: client,
+        let screenCoordinator = GiniNetworkingScreenAPICoordinator(client: client,
                                                          resultsDelegate: resultsDelegate,
                                                          giniConfiguration: configuration,
                                                          publicKeyPinningConfig: publicKeyPinningConfig,
